Refactor: moved Freshdesk models into tracking app.
